added Card Schema
